WFLY-14504 Refactor DefaultNonBlockingThreadFactory into wildfly-clustering-infinispan-spi
