+++ clustering/infinispan/extension/src/main/java/org/jboss/as/clustering/infinispan/subsystem/remote/ClientThreadPoolServiceConfigurator.java
@@ -40,13 +40,13 @@ import org.infinispan.commons.executors.ExecutorFactory;
 import org.infinispan.commons.util.concurrent.BlockingRejectedExecutionHandler;
 import org.infinispan.commons.util.concurrent.NonBlockingRejectedExecutionHandler;
 import org.jboss.as.clustering.context.DefaultThreadFactory;
-import org.jboss.as.clustering.infinispan.DefaultNonBlockingThreadFactory;
 import org.jboss.as.clustering.infinispan.subsystem.ComponentServiceConfigurator;
 import org.jboss.as.clustering.infinispan.subsystem.ThreadPoolDefinition;
 import org.jboss.as.controller.OperationContext;
 import org.jboss.as.controller.OperationFailedException;
 import org.jboss.as.controller.PathAddress;
 import org.jboss.dmr.ModelNode;
+import org.wildfly.clustering.infinispan.spi.DefaultNonBlockingThreadFactory;
 import org.wildfly.clustering.service.ServiceConfigurator;
 
 /**
